Removed uses of eval and added try/catch in case of incorrect links for Link search
Previous code ran eval() on part of the Link search, allowing one to input text which would run potentially malicious (client-side) Javascript. In addition, a new try/catch function has been added, in case of invalid links which cannot be parsed, in which case a generic error message is posted.